+++ src/renderer/views/projects/details/LatestVersionCell.tsx
@@ -2,20 +2,74 @@ import React, { useCallback, useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import Loading from '@renderer/components/Loading/Loading';
 import ErrorIcon from '@renderer/components/Icon/ErrorIcon';
+import PackageVersionTag, {
+  PackageVersionTagColor,
+} from '@renderer/components/Tag/Tag';
+import i18n from '../../../i18n';
 
 interface Props {
   dependencyName: string;
+  dependencyCurrenVersion: string;
   registryUrl?: string;
 }
 
+const computeTagColor = (
+  currentVersion: string,
+  latestVersion?: string,
+): PackageVersionTagColor | undefined => {
+  if (!latestVersion) {
+    return undefined;
+  }
+
+  const splitCurrentVersion = currentVersion.split('.');
+  const splitLatestVersion = latestVersion.split('.');
+
+  if (splitCurrentVersion.length !== 3 || splitLatestVersion.length !== 3) {
+    return undefined;
+  }
+
+  if (splitCurrentVersion[0] < splitLatestVersion[0]) {
+    return PackageVersionTagColor.RED;
+  }
+
+  if (splitCurrentVersion[1] < splitLatestVersion[1]) {
+    return PackageVersionTagColor.ORANGE;
+  }
+
+  if (splitCurrentVersion[2] < splitLatestVersion[2]) {
+    return PackageVersionTagColor.BLUE;
+  }
+
+  return undefined;
+};
+
+const computeTagTooltip = (
+  tagColor?: PackageVersionTagColor,
+): string | undefined => {
+  switch (tagColor) {
+    case PackageVersionTagColor.RED:
+      return i18n.t('project.details.tooltip.newMajor');
+    case PackageVersionTagColor.ORANGE:
+      return i18n.t('project.details.tooltip.newMinor');
+    case PackageVersionTagColor.BLUE:
+      return i18n.t('project.details.tooltip.newPatch');
+    default:
+      return undefined;
+  }
+};
+
 const LatestVersionCell: React.FunctionComponent<Props> = (props) => {
-  const { dependencyName, registryUrl } = props;
+  const { dependencyName, dependencyCurrenVersion, registryUrl } = props;
 
   const { t } = useTranslation();
 
-  const [isLoading, setIsLoading] = useState<boolean>(false);
+  const [isLoading, setIsLoading] = useState<boolean>(true);
 
   const [latestVersion, setLatestVersion] = useState<string | undefined>();
+
+  const [versionTagColor, setVersionTagColor] = useState<
+    PackageVersionTagColor | undefined
+  >(undefined);
 
   const fetchLatestVersion = useCallback(() => {
     setIsLoading(true);
@@ -27,11 +81,12 @@ const LatestVersionCell: React.FunctionComponent<Props> = (props) => {
       .then((result) => {
         setIsLoading(false);
         setLatestVersion(result);
+        setVersionTagColor(computeTagColor(dependencyCurrenVersion, result));
       })
       .catch(() => {
         setIsLoading(false);
       });
-  }, [dependencyName, registryUrl]);
+  }, [dependencyName, registryUrl, dependencyCurrenVersion]);
 
   useEffect(() => {
     fetchLatestVersion();
@@ -42,7 +97,11 @@ const LatestVersionCell: React.FunctionComponent<Props> = (props) => {
       {isLoading ? (
         <Loading className="cell-loading" />
       ) : latestVersion ? (
-        <p>{latestVersion}</p>
+        <PackageVersionTag
+          content={latestVersion}
+          color={versionTagColor}
+          tooltip={computeTagTooltip(versionTagColor)}
+        />
       ) : (
         <ErrorIcon tooltip={t('project.details.table.values.unableToFetch')} />
       )}